Added simple game logic to demo
Implement a simple logic to the game. Rather than instantly moving the
square, we will now instead give a command to the object. The deltaX
and deltaY would record the x and y movement for the square. Their
value would be calculated from the speed of the object and the angle
they are traveling at.